Add support for AWS profile in remote cache DSL
This allows the remote cache configuration to use a profile name when
obtaining AWS credentials. Profiles are an alternate way to provide
credentials. The current method of doing this is to set
`lookupDefaultAwsCredentials=true` and the environment variable
`export AWS_PROFILE="<profile_name>"`, but this requires each developer
to set that variable. Moving it into the remote cache configuration
allows that setting to be more easily shared, particularly in
environments that used shared managed profiles.

More about using profiles at
https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/credentials-profiles.html